# 터레인맵 <->json맵 합체중

+++ Code/Majestro/Majestro_Engine/TerrainComponent.h
@@ -18,7 +18,7 @@ class TerrainComponent : public Component<TerrainComponent>
 {
 public:
 		TerrainComponent() {}
-		TerrainComponent(float width, float depth, shared_ptr< Material>& material);
+		TerrainComponent(int width, int depth, shared_ptr< Material>& material);
 
 
 		float GetHeightAtWorldPosition(const Vec3& worldPos) const;
